ai_content_intelligence/requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# API clients
openai>=1.0.0
anthropic>=0.18.0
httpx>=0.25.0

# Utility
requests>=2.31.0
pillow>=10.0.0

# Database
supabase>=2.0.0
